fix ingress resources and tls when using --ingress-host-name
When using the --ingress-host-name flag we end up needing two separate
ingresses per service. This is because the cluster services will
continue to use the "host" domain to access other resources within the
cluster. In order to properly setup these extra ingresses we need to add
a SAN for the separate ingress host name.

Signed-off-by: Caleb Boylan <calebboylan@gmail.com>